# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elegant Sip API",
    "version": "0.1.0",
    "description": "Backend for the Elegant Sip storefront — single-origin Darjeeling tea.\n\n**Money is integer paise** (1 rupee = 100 paise). Razorpay settles in paise and\nfloats cannot represent money. Totals are always a whole number of rupees:\nGST is rounded with the same rule the storefront uses, so the figure a customer\nis shown and the figure they are charged cannot drift apart.\n\nErrors share one shape across every endpoint, modelled on RFC 9457. `detail` is\nalways safe to display; `requestId` ties a response to the server logs.",
    "contact": {
      "name": "Elegant Sip",
      "email": "elegantsipdarjeeling@gmail.com"
    }
  },
...
  "paths": {
...
    "/v1/admin/orders": {
      "get": {
        "summary": "The order queue",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending_payment",
                "paid",
                "packed",
                "shipped",
                "delivered",
                "cancelled",
                "refunded"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "number": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "placedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          "email": {
                            "type": "string"
                          },
                          "total": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991,
                            "description": "Amount in paise (1 rupee = 100 paise)"
                          },
                          "itemCount": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "shippingCity": {
                            "type": "string"
                          },
                          "trackingNumber": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "number",
                          "status",
                          "placedAt",
                          "email",
                          "total",
                          "itemCount",
                          "shippingCity",
                          "trackingNumber"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "orders"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders/{number}": {
      "patch": {
        "summary": "Advance an order",
        "tags": [
          "Admin"
        ],
        "description": "Setting status to `shipped` with a tracking number emails the customer. Nothing is emailed unless the tracking details are actually present — the message would otherwise promise information it does not carry.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "packed",
                      "shipped",
                      "delivered",
                      "cancelled"
                    ]
                  },
                  "trackingCarrier": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "trackingNumber": {
                    "type": "string",
                    "maxLength": 60
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 40
            },
            "in": "path",
            "name": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "notified": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok",
                    "notified"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders/{number}/invoice": {
      "get": {
        "summary": "GST invoice for an order",
        "tags": [
          "Admin"
        ],
        "description": "Returns the invoice as structured data. Without SELLER_GSTIN set, `isTaxInvoice` is false and the document is provisional — it cannot claim to be a tax invoice without a GSTIN.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 40
            },
            "in": "path",
            "name": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "invoice": {}
                  },
                  "required": [
                    "invoice"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/stock": {
      "post": {
        "summary": "Adjust stock",
        "tags": [
          "Admin"
        ],
        "description": "Applies a delta and records it in the append-only ledger, so the running total always has an explanation behind it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productSlug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                  },
                  "variantSize": {
                    "type": "string",
                    "minLength": 1
                  },
                  "delta": {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "productSlug",
                  "variantSize",
                  "delta"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "productSlug": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                      "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                    },
                    "variantSize": {
                      "type": "string"
                    },
                    "stock": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "productSlug",
                    "variantSize",
                    "stock"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.elegantsip.com",
      "description": "Production"
    }
  ],
...
}